api/Marketing_Expert/requirements.txt
langchain_community
langchain-google-genai
langchain_openai
streamlit==1.31.1
langchain_core
python-dotenv
langserve
fastapi
uvicorn
sse_starlette
langchainhub
langchain-groq
groq